Changment de façon de lire les fichiers (utilisation de Scanner)

+++ src/map2/Wall.java
@@ -2,10 +2,10 @@ package map2;
 
 public class Wall {
 	
-	private int x;
-	private int y;
-	private int thickness;
-	public Wall(int x,int y, int thickness){
+	private float x;
+	private float y;
+	private float thickness;
+	public Wall(float x,float y, float thickness){
 		this.x=x;
 		this.y = y;
 		this.thickness = thickness;
